requirements.txt
dash==1.16.3
dash-renderer==1.8.2
dash-html-components==1.1.1
dash-core-components==1.12.1
gunicorn>=20.1.0
pandas>=1.3.2
plotly>=4.11.0
python-aqi>=0.6.1
waqi-python>=0.2.2
